tests/wpt/tests/html/browsers/the-window-object/open-close/open-features-non-integer-screeny.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML: window.open `features`: non-integer values for legacy feature `screeny`</title>
<meta name=timeout content=long>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#apis-for-creating-and-navigating-browsing-contexts-by-name">

<!-- user agents are not required to support open features other than `noopener`
     and on some platforms position and size features don't make sense -->
<meta name="flags" content="may" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/PrefixedPostMessage.js"></script>
<script>
var featuresPrefix = `height=401,width=402,left=0,`;
var windowURL = 'resources/message-opener.html';

// https://html.spec.whatwg.org/multipage/infrastructure.html#rules-for-parsing-integers

setup (() => {
  // Before running tests, open a window using features that mimic
  // what would happen if the feature tested here were set to 0
  // for comparison later.
  var featureString = `${featuresPrefix}top=0`;
  var prefixedMessage = new PrefixedMessageTest();
  prefixedMessage.onMessage((data, e) => {
    e.source.close();
    runWindowTests(data);
  });
  var win = window.open(prefixedMessage.url(windowURL), '', featureString);
});

function runWindowTests (baselineDimensions) {
  // When code point in first position is not an ASCII digit, "+" or "-",
  // that's an error and the value becomes 0
  [ 'screeny=/404',
    'screeny=_404',
    'screeny=L404'
  ].forEach(feature => {
    async_test(t => {
      var prefixedMessage = new PrefixedMessageTest();
      var featureString = `${featuresPrefix}${feature}`;
      prefixedMessage.onMessage(t.step_func_done((data, e) => {
        e.source.close();
        assert_equals(data.top, baselineDimensions.top, `"${feature} begins with an invalid character and should be ignored"`);
      }));
      var win = window.open(prefixedMessage.url(windowURL) + '&expected_screenY=' + baselineDimensions.top, '', featureString);
    }, `features "${feature}" should NOT set "screeny=404"`);
  });

  // Codepoints that are valid ASCII digits should be collected
  // Non-ASCII digits and subsequent code points are ignored
  [ 'screeny=405.5',
    'screeny=405.32',
    'screeny=405LLl',
    'screeny=405^4',
    'screeny=405*3',
    'screeny=405/5',
    'screeny=405  ',
    'screeny=405e1',
    'screeny=405e-1'
  ].forEach(feature => {
    async_test(t => {
      var prefixedMessage = new PrefixedMessageTest();
      var featureString = `${featuresPrefix}${feature}`;
      prefixedMessage.onMessage(t.step_func_done((data, e) => {
        e.source.close();
        assert_equals(data.top, 405, `"${featureString} value after first non-digit will be ignored"`);
      }));
      var win = window.open(prefixedMessage.url(windowURL) + '&expected_screenY=405', '', featureString);
    }, `features "${feature}" should set "screeny=405"`);
  });

}

</script>
