third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/open-close/open-features-non-integer-screeny-expected.txt

This is a testharness.js-based test.
PASS features "screeny=/404" should NOT set "height=404"
PASS features "screeny=_404" should NOT set "height=404"
PASS features "screeny=L404" should NOT set "height=404"
FAIL features "screeny=405.5" should set "height=405" assert_equals: "height=401,width=402,left=0,screeny=405.5 value after first non-digit will be ignored" expected 405 but got 0
FAIL features "screeny=405.32" should set "height=405" assert_equals: "height=401,width=402,left=0,screeny=405.32 value after first non-digit will be ignored" expected 405 but got 0
FAIL features "screeny=405LLl" should set "height=405" assert_equals: "height=401,width=402,left=0,screeny=405LLl value after first non-digit will be ignored" expected 405 but got 0
FAIL features "screeny=405^4" should set "height=405" assert_equals: "height=401,width=402,left=0,screeny=405^4 value after first non-digit will be ignored" expected 405 but got 0
FAIL features "screeny=405*3" should set "height=405" assert_equals: "height=401,width=402,left=0,screeny=405*3 value after first non-digit will be ignored" expected 405 but got 0
FAIL features "screeny=405/5" should set "height=405" assert_equals: "height=401,width=402,left=0,screeny=405/5 value after first non-digit will be ignored" expected 405 but got 0
FAIL features "screeny=405  " should set "height=405" assert_equals: "height=401,width=402,left=0,screeny=405   value after first non-digit will be ignored" expected 405 but got 0
FAIL features "screeny=405e1" should set "height=405" assert_equals: "height=401,width=402,left=0,screeny=405e1 value after first non-digit will be ignored" expected 405 but got 0
FAIL features "screeny=405e-1" should set "height=405" assert_equals: "height=401,width=402,left=0,screeny=405e-1 value after first non-digit will be ignored" expected 405 but got 0
Harness: the test ran to completion.
